requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
inflection==0.5.1
packaging==23.1
phonenumbers==8.13.16
Pillow==10.0.0
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
